+++ packages/vue-cli-plugin-vue-dash/generator/template/src/views/tests/NotFound.spec.ts
@@ -1,16 +1,34 @@
 import Vue from 'vue';
-import { Wrapper } from '@vue/test-utils';
+import Vuetify from 'vuetify';
 
-import { mountComponent } from '@/tests-unit';
-import { html } from '@cnamts/vue-dot/tests/utils/html';
+import {
+	Wrapper,
+	html,
+	shallowMount,
+	createLocalVue,
+	createVuetifyInstance,
+	installGlobalPlugins
+} from '@cnamts/vue-dot/src/helpers/testUtils';
 
 import NotFound from '../NotFound.vue';
 
-let wrapper: Wrapper<Vue>;
-
 describe('NotFound', () => {
+	const localVue = createLocalVue();
+
+	let wrapper: Wrapper<Vue>;
+	let vuetify: Vuetify;
+
+	installGlobalPlugins(localVue);
+
+	beforeEach(() => {
+		vuetify = createVuetifyInstance();
+	});
+
 	it('renders correctly', () => {
-		wrapper = mountComponent(NotFound);
+		wrapper = shallowMount(NotFound, {
+			localVue,
+			vuetify
+		});
 
 		expect(html(wrapper)).toMatchSnapshot();
 	});