+++ packages/vue-cli-plugin-vue-dash/generator/template/tests/unit/setup.ts
@@ -1,0 +1,3 @@
+import { applyGlobalConfig } from '@cnamts/vue-dot/src/helpers/testUtils';
+
+applyGlobalConfig();